options refactor
